requirements.txt
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
tabulate>=0.9.0
